pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["NeSy4PPM", "NeSy4PPM.*"]


[project]
name = "nesy4ppm"
version = "0.1.0"
#dynamic = ["version"]
description = "Python library to perform learning and suffix prediction with background knowledge under concept drifts."
readme = "README.md"
requires-python = ">=3.9.7"
#license = {text = "GPL"}
keywords = ["python", "process mining", "predictive process monitoring", "suffix prediction", "background knowledge"]
authors = [
    { name = "Jamila Oukharijane", email = "jamila.oukharijane@unibz.it" },
    { name = "Ivan Donadello", email = "ivan.donadello@unibz.it" },
    { name = "Fabrizio Maria Maggi", email = "maggi@inf.unibz.it" }
]

dependencies = [
    "declare4py>=2.2.0",
    "tensorflow==2.10.1",
    "Distance==0.1.3",
    "keras==2.10.0",
    "keras-nlp==0.4.1",
    "jellyfish==0.9.0",
    "absl-py==2.3.0"

]

[project.urls]
Homepage = "https://github.com/JamilaOUKHARIJANE/NeSy4PPM"


